Commit to move JWT AND password generation files to utilities folder
maintain(controllers/password, controllers/jwt-token): move files to utilities folder
- move password and jwt-token files to utilities folder
- modify path to the moved files in all dependent modules
- [Partly Delivers #165437339]
