+++ utilities/password.js
